+++ CHANGELOG.md
@@ -4,10 +4,10 @@
 
 Released on 
 
-- `check library` command now ignores the `.git` folder.
+- `check library` command now ignores the `.git` and `.github` folders, and `.gitignore` file.
 - Modify html output of `html` command.
     - Margins are smaller in mobile devices.
-    - Can now show notes by folder.
+    - Can now show entries by folder.
 - In the TUI, the editor for the entry and note is taken from the `EDITOR` environment variable. If not set, it raises a warning.
 
 Pypi: https://pypi.org/project/bibmancli/0.3.2/